pyproject.toml
[tool.poetry]
name = "automation-assistant"
version = "0.1.0"
description = "Automation workflow generator using LLM and n8n"
authors = ["Artem Shishkin <verter25@gmail.com>"]
readme = "README.md"

packages = [
  { include = "automation_assistant" }
]

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.88.0"
requests = "^2.32.4"
python-dotenv = "^1.1.0"
jsonschema = "^4.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
guardrails-ai = "^0.6.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
